When measuring elapsed time for blockingtimeout, use a monotonic clock, rather than the system clock.
This will make it more tolerant of clock changes due to ntp or VM clock
synchronization, particularly on misconfigured collectors with incorrect
vmware host times and sync enabled.

Note that this change does *not* address similar issues that could occur with
timeouts that are implemented with twisted.  This includes the running timeout
and various API timeouts in other zenpacks.  However, the stakes of false
alarms on those timeouts are significantly lower, compared to runningtimeout,
which will disable zenpython plugins indefinitely if it trips.

There is an extensive discussion of the twisted issue at https://twistedmatrix.com/trac/ticket/2424,
but no clear decision seems to have been made at this time.